feat(epicenter): give Home three panes over the application catalog
Home was one chat screen whose header had become the place every control landed:
a tool popover, New chat, and a Local model toggle that expanded administration
inline above the conversation. Nothing listed applications at all.

Apps, Chat, and Settings are now siblings. Apps renders the host's one list with
one row and one action; Chat keeps the session and takes back its own controls;
Settings owns model administration. Tabs.Content hides an inactive pane instead
of unmounting it and the session is owned above them, so switching panes leaves
the socket, transcript, and an unsent draft alone.

Which pane Home lands on is decided synchronously from the injected host
bootstrap, so there is no first paint to correct. A browser or remote Home says
where applications and device-local settings live rather than offering buttons
that would fail, and open_home(transcription) selects Settings with the same
take-once intent it always had.
